% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_city_coord.R
\name{get_city_coord}
\alias{get_city_coord}
\title{Get the location information of the city}
\usage{
get_city_coord(city)
}
\arguments{
\item{city}{a character object of the name of Chinese city,}
}
\value{
a vecctor of longtitude and latitude
}
\description{
get_city_coord use Baidu API  <\url{http://developer.baidu.com/map/}>
to get the longtitude and latitude of the city
}
\details{
This function can transform the name of Chinese city Which
can be a Chinese string or an English name,like Shanghai.
This function is used BaiduAPI to get the geo information.
}
\examples{
get_city_coord("北京")
get_city_coord("Shanghai")
}
\author{
Chiffon <\url{http://chiffon.gitcafe.io}>
}

