% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_geo_position.R
\name{get_geo_position}
\alias{get_geo_position}
\title{Get the location information of the city vector}
\usage{
get_geo_position(city_vec)
}
\arguments{
\item{city_vec}{a character object of the name of Chinese city,}
}
\value{
A dataframe of longtitude and latitude and city name
}
\description{
get_geo_position use the function of get_city_coord according
to Baidu API <\url{http://developer.baidu.com/map/}>to get the longtitude and latitude of the city
vector
}
\details{
This function can transform the name of Chinese city ve tor
Which can be a Chinese string or an English name,like Shanghai.
This function is used BaiduAPI to get the geo information.\\
It's recomendece to use Chinese city name because of the API
we use is Baidu, which will performs better in Chiese\\
Please make sure your bowser can open developer.baidu.com/map/ to get the data.
}
\examples{
city_vec1 = c("Beijing","Shanghai","Guangzhou")
city_vec2 = c("北京","上海","广州")
get_geo_position (city_vec1)
get_geo_position (city_vec2)
}
\author{
Dawei.Lang Chiffon <\url{http://chiffon.gitcafe.io}>
}

