% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_theme.R
\name{get_theme}
\alias{get_theme}
\title{Create a theme object for remap}
\usage{
get_theme(theme = "Dark", lineColor = "Random",
  backgroundColor = "#1b1b1b", titleColor = "#fff",
  borderColor = "rgba(100,149,237,1)", regionColor = "#1b1b1b")
}
\arguments{
\item{theme}{a character object in ("Dark","Bright,"Sky","none)}

\item{lineColor}{Control the color of the line, "Random" for
random color}

\item{backgroundColor}{Control the color of the background}

\item{titleColor}{Control the color of the title}

\item{borderColor}{Control the color of the border}

\item{regionColor}{Control the color of the region}
}
\value{
A list of color control, which can be used by remap
}
\description{
get_theme create a theme list to control the color
we see in the remap.including lineColor, backgroundColor,
titleColor, borderColor and regionColor.
}
\details{
If you use the theme argument of the get_theme function,
you will get the default theme in one of ("Dark","Bright,"Sky").
If you don't like the color, set theme = "none" and use other
parameters control remap.\\
Can use "red","#1b1b1b" or rgba(100,100,100,1) to control the color
}
\examples{
## default theme:"Dark"
set.seed(125)
out = remap(demoC,title = "REmap示例数据",subtitle = "theme:Bright",
            theme = get_theme("Bright"))
plot(out)

## set Line color as 'orange'
set.seed(125)
out = remap(demoC,title = "REmap示例数据",subtitle = "theme:Bright",
            theme = get_theme("None",
                             lineColor = "orange"))
plot(out)

## Set backgroundColor as 'red'(#FF0000)

out = remap(demoC,title = "REmap示例数据",subtitle = "theme:Bright",
            theme = get_theme("None",
                              lineColor = "orange",
                              backgroundColor = "#FF0000"))
plot(out)

## Set TitleColor
out = remap(demoC,title = "REmap示例数据",subtitle = "theme:Bright",
            theme = get_theme("None",
                              lineColor = "orange",
                              backgroundColor = "#FFC1C1",
                              titleColor = "#1b1b1b"))
plot(out)

## Set Region Color
out = remap(demoC,title = "REmap示例数据",subtitle = "theme:Bright",
            theme = get_theme("None",
                              lineColor = "orange",
                              backgroundColor = "#FFC1C1",
                              titleColor = "#1b1b1b",
                              regionColor = '#ADD8E6'))
plot(out)
}
\author{
Chiffon <\url{http://chiffon.gitcafe.io}>
}

