% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/knitrREmap.R
\name{knitrREmap}
\alias{knitrREmap}
\title{Plot REmap in knitr}
\usage{
knitrREmap(object, height = "300px", width = "100\%", local = F)
}
\arguments{
\item{height}{height for the DOM of REmap}

\item{width}{width for the DOM of REmap}

\item{objct}{an REmap object for plotting}
}
\description{
Function for plotting REmap in knitr
}
\details{
After using the remap.init function, you can use this function
for plotting REmap in knitr in any chuck
}
\examples{
library(REmap)
out = remap(demoC)
knitrREmap(out)
}

