% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mapNames.R
\name{mapNames}
\alias{mapNames}
\title{Get the names of a choropleth map for remapC}
\usage{
mapNames(mapType)
}
\arguments{
\item{mapType}{one of "china", "world" or the name of
province in China}
}
\value{
mapNames   returns a vec of map names
}
\description{
mapNames show what kind of map names is needed for remapC
}
\examples{
## get the province names:
mapNames('china')
}

