% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_rras_aps2_server.R
\name{mod_rras_aps2_server}
\alias{mod_rras_aps2_server}
\title{RRAS APS 2 Server}
\usage{
mod_rras_aps2_server(id, data_list)
}
\arguments{
\item{id}{Module id}

\item{data_list}{Lista com os dados carregados em load_data()}
}
\description{
RRAS APS 2 Server
}
