% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_rras_aps_server.R
\name{mod_rras_aps_server}
\alias{mod_rras_aps_server}
\title{RRAS APS Server}
\usage{
mod_rras_aps_server(id, data_list)
}
\arguments{
\item{id}{Module id}

\item{data_list}{Lista com os dados carregados em load_data()}
}
\description{
RRAS APS Server
}
