% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remap.R
\name{plot.remap}
\alias{plot.remap}
\title{Plot method for remap object.}
\usage{
\method{plot}{remap}(x, ...)
}
\arguments{
\item{x}{\emph{} S3 object output from remap.}

\item{...}{Arguments to pass to regions plot.}
}
\value{
A list that plots a map of the regions used for modeling.
}
\description{
Plots the regions used for modeling.
}
