% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remap.R
\name{print.remap}
\alias{print.remap}
\title{Print method for remap object.}
\usage{
\method{print}{remap}(x, ...)
}
\arguments{
\item{x}{\emph{} S3 object output from remap.}

\item{...}{Extra arguments.}
}
\value{
No return value, a description of the remap object is printed in the
console.
}
\description{
Print method for remap object.
}
