% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redist.R
\name{redist}
\alias{redist}
\title{Get distances between data and regions.}
\usage{
redist(
  data,
  lon,
  lat,
  regions,
  max_dist = NULL,
  region_id = NULL,
  progress = TRUE
)
}
\arguments{
\item{data}{An sf data frame with longitude, latitude, and columns required
for modeling.}

\item{lon}{Name of longitude column (no quotes).}

\item{lat}{Name of latitude column (no quotes).}

\item{regions}{An sf dataframe with polygon geometry.}

\item{max_dist}{a maximum distance that is needed for future calculations.
(Set equal to maximum 'smooth' when predicting on new observations.)}

\item{region_id}{Optional name of column in 'regions' that contains the id
that each region belongs to (no quotes). If null, it will be assumed that
each polygon is its own region (no regions have more than one polygon).}

\item{progress}{If true, a text progress bar is printed to the console.}
}
\value{
A matrix where each row corresponds one-to-one with each row in
provided 'data'. Matrix columns are either named with regions from 'region_id'
column of 'regions' or the row numbers of 'regions' if 'region_id' is NULL.
Values are in kilometers.
}
\description{
Finds distances in km between data provided as sf dataframe with point geometry
and regions provided as sf dataframe with polygon geometry.
}
\seealso{
\code{\link{remap}} - uses redist for regional models.
}
