% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/remap.init.R
\name{remap.init}
\alias{remap.init}
\title{Remap initial for knitr}
\usage{
recharts.init()
}
\description{
An shell function for initializing knitr for remap packages
}
\details{
Only the first chuck of recharts plot needs this function,
 the rest chuck mustn't include this function.
}

