% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remap.R
\name{remap}
\alias{remap}
\title{Build separate models for mapping multiple regions.}
\usage{
remap(
  data,
  model_function,
  buffer,
  min_n = 0,
  regions,
  region_id,
  distances,
  progress = FALSE,
  keep_data = FALSE,
  ...
)
}
\arguments{
\item{data}{An sf data frame with point geometry.}

\item{model_function}{A function that can take a subset of 'data' and
output a model that can be used to predict new values when passed to generic
function predict().}

\item{buffer}{A buffer zone around each region in km where data is included
in the data used to build models for each region. (Can be a named vector
with different values for each unique 'region_id' in 'region'.)}

\item{min_n}{The minimum number of observations to use when building a model.
If there are not enough observations in the region and buffer, then the
closest min_n observations are used. No minimum if set to 0.}

\item{regions}{An sf dataframe with polygon or multipolygon geometry.}

\item{region_id}{Optional name of column in 'regions' that contains the id
that each region belongs to (no quotes). If null, it will be assumed that
each polygon is its own region (no regions have more than one polygon).}

\item{distances}{An optional matrix of distances between 'data' and 'regions'
generated by redist() function (calculated internally if not
provided).}

\item{progress}{If true, a text progress bar is printed to the console.}

\item{keep_data}{If true, extra data is included in the output object. Extra
data is 'data', 'distances', and 'data_indices', (see output description).}

\item{...}{Extra arguments to pass to 'model_function' function.}
}
\value{
A \emph{remap} S3 object containing:
\describe{
  \item{\emph{models}}{A list of models containing a model output by
  'model_function' for each region.}
  \item{\emph{regions}}{'regions' object passed to the function (used for
  prediction).}
  \item{\emph{region_id}}{'region_id' object passed to the function (used for
  prediction).}
  \item{\emph{call}}{Shows the parameters that were passed to the function.}
  \item{\emph{data}}{'data' object passed to the function
  ('keep_data' = TRUE).}
  \item{\emph{distances}}{Distances returned by internal call of
  redist() function ('keep_data' = TRUE).}
  \item{\emph{data_indices}}{A list of indices of rows of 'data' that were
  used to build each regional model ('keep_data' = TRUE).}
}
}
\description{
Separate models are built for each given region and combined into one S3
object that can be used to predict on new data using generic function
predict(). If a model fails for a region, a warning is given but the modeling
process will continue.
}
\seealso{
\code{\link{predict.remap}} - used for predicting on new data.
  \code{\link{redist}} - used for building models within a buffer.
}
