% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/remapB.R
\name{remapB}
\alias{remapB}
\title{Create a Bmap object}
\usage{
remapB(center = c(104.114129,37.550339),
       zoom = 5,
       color = "Bright",
       title = "",
       subtitle = "",
       markLineData = NA,
       markPointData = NA,
       markLineTheme = markLineControl(),
       markPointTheme = markPointControl(),
       geoData = NA)
}
\arguments{
\item{center}{the center of the Bmap,a vector of (lon,lat),
you can get from get_city_coord, e.g.get_city_coord("beijing")}

\item{zoom}{the size of the Bmap, zoom:5 country data,
zoom:15 city data}

\item{color}{"Bright" or "Blue"}

\item{title}{a character string of the title}

\item{subtitle}{a character string of the subtitle}

\item{markLineData}{data for mark line}

\item{markPointData}{data for mark point}

\item{markLineTheme}{theme for mark line}

\item{geoData}{geoData for markLine and markPoint format is
similar as get_geo_position()}

\item{markLineTheme}{theme for mark point}
}
\value{
An remap object
}
\description{
remapB uses BaiduAPI instead of SVG object as the map object.
You can mark line or point on the remapB.
}
\examples{
geoData  = get_geo_position(unique(demoC[demoC==demoC]))
# this may take some time,be patient~

remapB(markLineData = demoC,geoData = geoData)



remapB(markLineData = demoC,color = "Blue",geoData = geoData)


remapB(markLineData = demoC,markPointData = demoC[,2],
       color = "Blue",geoData = geoData)
}
\author{
Chiffon <\url{http://chiffon.gitcafe.io}>
}

