% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/remapC.R
\name{remapC}
\alias{remapC}
\title{Create a choropleth map object}
\usage{
remapC(data,
       maptype = 'china',
      markLineData = NA,
      markPointData = NA,
      color = c('#1e90ff','#f0ffff'),
      theme = get_theme("Bright"),
      title = "",
      subtitle = "",
      markLineTheme = markLineControl(),
      markPointTheme = markPointControl(),
      geoData = NA,
      mindata = NA,
      maxdata = NA)
}
\arguments{
\item{data}{a data frame including place names and values}

\item{maptype}{the type of the map. For example,'china',
'world' or other names of province in China.}

\item{markLineData}{data for mark line}

\item{markPointData}{data for mark point}

\item{color}{vector of the color, if the length of color is
1, white will be added to the color.}

\item{theme}{a list object created by get_theme,control
the color of the map.}

\item{title}{a character string of the title}

\item{subtitle}{a character string of the subtitle}

\item{markLineTheme}{theme for mark line}

\item{geoData}{geoData for markLine and markPoint format is
similar as get_geo_position()}

\item{mindata}{mindata for legend}

\item{maxdata}{maxdata for legend}

\item{markLineTheme}{theme for mark point}
}
\value{
An remap object
}
\description{
remapC uses a data frame create a remap object
which can be used by plot to see the map by browser
}
\details{
mapdata should be a dataframe which including two columns,
the first column is the names of places,
second column is value.
}
\examples{
data = data.frame(country = mapNames("world"),
                   value = 5*sample(178)+200)
head(data)
out = remapC(data,maptype = "world",color = 'skyblue')
plot(out)

remapC(chinaIphone,
        markLineData = demoC,
        markPointData = demoC[,2])
}
\author{
Chiffon <\url{http://chiffon.gitcafe.io}>
}

