% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/renderREmap.R
\name{REmapOutput}
\alias{REmapOutput}
\title{Plot REmap in shiny}
\usage{
REmapOutput(outputId, inline = FALSE, container,...)
renderREmap(expr, env = parent.frame(),height = "800px",
           width = "100\%", quoted = FALSE, func = NULL)
}
\description{
Function for plotting REmap in shiny
}
\details{
USe renderREmap render an REmap object and use REmapOutput
output an REmap object. See more details in shiny package.
}
\examples{
library(REmap)
library(shiny)
runApp(system.file("shiny","shinyBmap", package = "REmap"))
}

