% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remap.R
\name{summary.remap}
\alias{summary.remap}
\title{Summary method for remap object.}
\usage{
\method{summary}{remap}(object, ...)
}
\arguments{
\item{object}{\emph{} S3 object output from remap.}

\item{...}{Extra arguments to pass to regional models.}
}
\value{
No return value, a brief summary of the remap object is printed in
the console. This includes the class(es) of the regional models, the
CRS of the regions, and the bounding box of the regions.
}
\description{
Summary method for remap object.
}
