% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{utsnow}
\alias{utsnow}
\title{Snowpack at weather stations in Utah on April 1st, 2011.}
\format{
An sf points object with 394 rows and 8 variables:
\describe{
  \item{ID}{Weather station identification code.}
  \item{STATION_NAME}{Weather station name.}
  \item{LATITUDE}{Latitude of weather station.}
  \item{LONGITUDE}{Longitude of weather station.}
  \item{ELEVATION}{Elevation of weather station.}
  \item{HUC2}{Largest watershed region containing this weather station
  (see \code{\link{utws}} data).}
  \item{WESD}{Water equivalent of snow density in mm of water.}
  \item{geometry}{sfc points in geographic coordinates.}
}
}
\source{
\url{https://www1.ncdc.noaa.gov/pub/data/ghcn/daily/}
}
\usage{
utsnow
}
\description{
Water equivalent of snow density (WESD) in mm of water at various
location within and surrounding the state of Utah. WESD are
measured at weather stations within the Daily Global Historical
Climatology Network. April first measurements are used to
estimate snowpack for the state of Utah.
}
\examples{
\donttest{
# If you run into issues with loading this dataset try running:
utsnow <- sf::st_read(system.file("extdata/utsnow.shp", package="remap"))
names(utsnow) <- c(
  "ID", "STATION_NAME", "LATITUDE", "LONGITUDE", "ELEVATION", "HUC2",
  "WESD", "geometry"
)
}
}
\keyword{datasets}
