% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.remap.R
\name{plot.remap}
\alias{plot.remap}
\title{plot the remap object}
\usage{
\method{plot}{remap}(object, path = "")
}
\arguments{
\item{object}{an remap object}

\item{path}{the path for saving REmap object}
}
\value{
Create a html file according to the id of the object
}
\description{
plot.remap is a function to create the htmlfile
and open it by browser
}
\examples{
set.seed(125)
out = remap(demoC,title = "REmap: Demo DATA",
subtitle = "theme:Dark")
plot(out)
}
\author{
Chiffon <\url{http://chiffon.gitcafe.io}>
}

