% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remap.R
\name{remap}
\alias{remap}
\title{Build separate models for mapping multiple regions.}
\usage{
remap(
  data,
  regions,
  region_id,
  model_function,
  buffer,
  min_n = 0,
  distances,
  cores = 1,
  progress = FALSE,
  ...
)
}
\arguments{
\item{data}{An sf data frame with point geometry.}

\item{regions}{An sf dataframe with polygon or multipolygon geometry.}

\item{region_id}{Optional name of column in 'regions' that contains the id
that each region belongs to (no quotes). If null, it will be assumed that
each polygon is its own region (no regions have more than one polygon).}

\item{model_function}{A function that can take a subset of 'data' and
output a model that can be used to predict new values when passed to generic
function predict().}

\item{buffer}{The length of the buffer zone around each region in km where
observations are included in the data used to build models for each region.
(Can be a named vector with different values for each unique 'region_id' in
'region'.)}

\item{min_n}{The minimum number of observations to use when building a model.
If there are not enough observations in the region and buffer, then the
closest min_n observations are used. No minimum if set to 0.}

\item{distances}{An optional matrix of distances between 'data' and 'regions'
generated by redist() function (calculated internally if not
provided). Note that unless you know that you have min_n within a certain
distance, no max_dist parameter should be used in redist().}

\item{cores}{Number of cores for parallel computing. 'cores' above
default of 1 will require more memory.}

\item{progress}{If true, a text progress bar is printed to the console.}

\item{...}{Extra arguments to pass to 'model_function' function.}
}
\value{
A \emph{remap} S3 object containing:
\describe{
  \item{\emph{models}}{A list of models containing a model output by
  'model_function' for each region.}
  \item{\emph{regions}}{'regions' object passed to the function (used for
  prediction).}
  \item{\emph{call}}{Shows the parameters that were passed to the function.}
}
}
\description{
Separate models are built for each given region and combined into one S3
object that can be used to predict on new data using generic function
predict(). If a model fails for a region, a warning is given but the modeling
process will continue.
}
\examples{
library(remap)
library(sf)
data(utsnow)
data(utws)

# Simplify polygons to run example faster
utws_simp <- sf::st_simplify(utws, dTolerance = 0.01)

# Build a remap model with lm that has formula snow_water = elevation
# The buffer to collect data around each region is 30km
# The minimum number of observations per region is 10
remap_model <- remap(data = utsnow,
                     regions = utws_simp,
                     region_id = HUC2,
                     model_function = lm,
                     formula = WESD ~ ELEVATION,
                     buffer = 20,
                     min_n = 10,
                     progress = TRUE)

# Resubstitution predictions
remap_preds <- predict(remap_model, utsnow, smooth = 10)
head(remap_preds)

}
\seealso{
\code{\link{predict.remap}} - used for predicting on new data.
  \code{\link{redist}} - used for building models within a buffer.
}
