% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/remap.r
\name{remap}
\alias{remap}
\title{Create an remap object}
\usage{
remap(mapdata, title = "", subtitle = "", theme = get_theme("Dark"))
}
\arguments{
\item{mapdata}{a data frame including depart and destination
in each column}

\item{title}{a character string of the title}

\item{subtitle}{a character string of the subtitle}

\item{theme}{a list object created by get_theme,control
the color of the map.}
}
\value{
An remap object
}
\description{
remap use a data frame create a remap object
which can be used by plot to see the map by browser
}
\details{
mapdata should be a dataframe which including two columns,
the first column is depart, second column is destination.
}
\examples{
set.seed(125)
out = remap(demoC,title = "REmap",subtitle = "theme:Dark")
plot(out)
summary(out)
}
\author{
Chiffon <\url{http://lchiffon.github.io}>
}

